+++ mockspresso-mockito-powermock/src/test/java/com/episode6/hackit/mockspresso/mockito/powermock/PowerMockitoConfigTest.java
@@ -24,18 +24,4 @@ public class PowerMockitoConfigTest {
             Spy.class,
             Mock.class);
   }
-
-  @Test
-  public void testMockMakerIsExpected() {
-    assertThat(mMockitoConfig.provideMockMaker())
-        .isNotNull()
-        .isInstanceOf(PowerMockitoMockMaker.class);
-  }
-
-  @Test
-  public void testFieldPreparerIsExpected() {
-    assertThat(mMockitoConfig.provideFieldPreparer())
-        .isNotNull()
-        .isInstanceOf(PowerMockitoFieldPreparer.class);
-  }
 }
