+++ mockspresso-mockito-powermock/src/test/java/com/episode6/hackit/mockspresso/mockito/powermock/PowerMockitoMockMakerTest.java
@@ -18,11 +18,11 @@ import static org.mockito.Mockito.verify;
 @RunWith(JUnit4.class)
 public class PowerMockitoMockMakerTest {
 
-  private PowerMockitoMockMaker mMockMaker;
+  private PowerMockitoConfig mMockMaker;
 
   @Before
   public void setup() {
-    mMockMaker = new PowerMockitoMockMaker();
+    mMockMaker = new PowerMockitoConfig();
   }
 
   @Test
